Move imports for the example plugin too.
