+++ src/sqlfluff/cli/helpers.py
@@ -2,9 +2,12 @@
 
 import sys
 import textwrap
-from typing import Dict, List
+from typing import Dict, List, Callable
+
+from collections import abc
 
 from sqlfluff import __version__ as pkg_version
+from sqlfluff.core.cached_property import cached_property
 
 
 def get_python_version() -> str:
@@ -72,3 +75,25 @@ def pad_line(s: str, width: int, align: str = "left") -> str:
         return (" " * gap) + s
     else:
         raise ValueError(f"Unknown alignment: {align}")  # pragma: no cover
+
+
+class LazySequence(abc.Sequence):
+    """A Sequence which only populates on the first access.
+
+    This is useful for being able to define sequences within
+    the click cli decorators, but that don't trigger their
+    contents until first called.
+    """
+
+    def __init__(self, getter=Callable[[], abc.Sequence]):
+        self._getter = getter
+
+    @cached_property
+    def _sequence(self) -> abc.Sequence:
+        return self._getter()
+
+    def __getitem__(self, key):
+        return self._sequence[key]
+
+    def __len__(self):
+        return len(self._sequence)
